domakin/requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==2.6.0
cryptography==44.0.0
Django==5.1.4
django-allauth==65.3.1
django-appconf==1.0.6
django-browser-reload==1.17.0
django-compressor==4.5.1
django-htmx==1.21.0
django-tailwind==3.8.0
idna==3.10
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
rcssmin==1.1.2
requests==2.32.3
rich==13.9.4
rjsmin==1.2.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
urllib3==2.3.0
